move swarming module to the swarming service
